Implement: Update kafka/message/fame/kafka.go - message struct fields [#255]
- Update Command struct WorldId field: byte → world.Id
- Update RequestChangeCommandBody struct fields:
  - ChannelId: byte → channel.Id
  - MapId: uint32 → _map.Id
- Update StatusEvent struct WorldId field: byte → world.Id
- Update StatusEventErrorBody struct ChannelId field: byte → channel.Id
- Add required imports for specialized identifier types
- Remove explicit type casts in fame/producer.go

This ensures Kafka fame messages use specialized identifier types
from atlas-constants instead of primitive types, improving type safety
and consistency across the messaging layer.

+++ atlas.com/channel/fame/producer.go
@@ -11,12 +11,12 @@ import (
 func RequestChangeFameCommandProvider(m _map.Model, characterId uint32, targetId uint32, amount int8) model.Provider[[]kafka.Message] {
 	key := producer.CreateKey(int(characterId))
 	value := &fame2.Command[fame2.RequestChangeCommandBody]{
-		WorldId:     byte(m.WorldId()),
+		WorldId:     m.WorldId(),
 		CharacterId: characterId,
 		Type:        fame2.CommandTypeRequestChange,
 		Body: fame2.RequestChangeCommandBody{
-			ChannelId: byte(m.ChannelId()),
-			MapId:     uint32(m.MapId()),
+			ChannelId: m.ChannelId(),
+			MapId:     m.MapId(),
 			TargetId:  targetId,
 			Amount:    amount,
 		},
